+++ transip/v6/services/vps.py
@@ -24,6 +24,11 @@ from transip.mixins import GetMixin, DeleteMixin, ListMixin
 from transip.v6.objects.vps import Vps
 
 
+class Vps(ApiObject):
+
+    _id_attr: str = "name"
+
+
 class VpsService(GetMixin, DeleteMixin, ListMixin, ApiService):
 
     _path: str = "/vps"
